Develop

+++ ecommerce.Core/Repositories/Interfaces/IRepository.cs
@@ -7,7 +7,7 @@ namespace ecommerce.Core.Repositories.Interfaces
     {
         public Task SaveChangesAsync();
         public Task<IEnumerable<TEntity>> GetAllAsync();
-        public Task<TEntity> GetByIdAsync(long id);
+        public Task<TEntity> GetByIdAsync(int id);
         public Task AddAsync(TEntity entity);
         public void Update(TEntity entity);
         public void Delete(TEntity entity);
